Update monitoring.rst
Having only one side hardly does the topic justice.  It's de rigeur to knock nagios these days, but most of the time, the arguments stem more from poor practice than actual deficiencies in the software. The *tools* section for Nagios should not consist solely of a naysayers opinion.  The first link is for @lozzd, who has contributed to the Nagios ecosystem, and the second is to an article by Dave Josephsen, who has (literally) written the book on building Enterprise grade monitoring systems with Nagios.